feat(client): exposet WithCapability type helper
