+++ packages/libs/client/src/composePactCommand/utils/addSigner.ts
@@ -1,4 +1,8 @@
 import { ICapabilityItem, IPactCommand } from '../../interfaces/IPactCommand';
+import {
+  ExtractCapabilityType,
+  IGeneralCapability,
+} from '../../interfaces/type-utilities';
 
 import { patchCommand } from './patchCommand';
 
@@ -59,25 +63,8 @@ export const addSigner: IAddSigner = ((
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 }) as any;
 
-export type UnionToIntersection<T> = (
-  T extends unknown ? (k: T) => void : never
-) extends (k: infer I) => void
-  ? I
-  : never;
-
-export interface IGeneralCapability {
-  (name: string, ...args: unknown[]): ICapabilityItem;
-  (name: 'coin.GAS'): ICapabilityItem;
-}
-
 type ExtractType<TCmdReducer> = TCmdReducer extends (cmd: {
   payload: infer TPayload;
 }) => unknown
-  ? TPayload extends { funs: infer TFunctions }
-    ? TFunctions extends Array<infer TFunction>
-      ? UnionToIntersection<TFunction> extends { capability: infer TCapability }
-        ? TCapability
-        : IGeneralCapability
-      : IGeneralCapability
-    : IGeneralCapability
+  ? ExtractCapabilityType<{ payload: TPayload }>
   : IGeneralCapability;
