+++ packages/libs/client/src/signing/ISignFunction.ts
@@ -2,9 +2,9 @@ import { ICommand, IUnsignedCommand } from '@kadena/types';
 
 export interface ISignFunction {
   (transaction: IUnsignedCommand): Promise<ICommand | IUnsignedCommand>;
-  (
-    transactionList: IUnsignedCommand[],
-  ): Promise<(ICommand | IUnsignedCommand)[]>;
+  (transactionList: IUnsignedCommand[]): Promise<
+    (ICommand | IUnsignedCommand)[]
+  >;
 }
 
 export interface ISignSingleFunction {
